requirements.txt
pip==21.3.1
setuptools==60.2.0
wheel==0.37.1
umls-api==0.1.0
psycopg==3.1.17